% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_grps.R
\name{add_grps}
\alias{add_grps}
\title{Collapse variable into groups and sum}
\usage{
add_grps(df, grp_list, group = group, estimate = estimate, moe = NULL)
}
\arguments{
\item{df}{A data frame; will honor grouping}

\item{grp_list}{A named list of groups to collapse \code{group} into}

\item{group}{Bare column name giving groups in data; will be converted to factor}

\item{estimate}{Bare column name of estimates or values}

\item{moe}{Bare column name of margins of error; if supplied, MOEs of sums will be included in output}
}
\value{
A data frame/tibble with sums of \code{estimate}. Retains grouping columns
}
\description{
Collapse variable into groups and sum
}
\examples{
unique(race_pops$variable)
race_list <- list(black_latino = c(3, 5), poc = 3:5)
race_pops \%>\%
  dplyr::group_by(region, name) \%>\%
  add_grps(race_list, group = variable, moe = moe)
}
