% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_shares.R
\name{calc_shares}
\alias{calc_shares}
\title{Make table of rates given a denominator}
\usage{
calc_shares(df, ..., group = variable, denom = "Total",
  estimate = estimate, moe = NULL)
}
\arguments{
\item{df}{A data frame}

\item{...}{Bare column names of IDs, town names, etc to exclude from \code{gather}ing}

\item{group}{Bare column name where groups are given}

\item{denom}{String; denominator to filter from `group``}

\item{estimate}{Bare column name of estimates or values}

\item{moe}{Bare column name of margins of error; if supplied, MOE of shares will be included in output}
}
\value{
A tibble/data frame with shares (and optionally MOE of shares) of subgroup values within a denominator group. Shares given for denominator group will be blank.
}
\description{
Make table of rates given a denominator
}
\examples{
edu \%>\%
  rename(est = estimate) \%>\%
  calc_shares(name, denom = "age25plus", estimate = est, moe = moe)
}
