% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cap_first.R
\name{cap_first}
\alias{cap_first}
\title{Capitalize first letter in string}
\usage{
cap_first(x)
}
\arguments{
\item{x}{A character vector}
}
\value{
A character vector with first character of each item capitalized
}
\description{
Capitalize first letter in string
}
\examples{
cap_first("camille")
cap_first(c("camille", "new haven"))
}
