% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_titles.R
\name{clean_titles}
\alias{clean_titles}
\title{Several options for formatting text into neater titles}
\usage{
clean_titles(
  x,
  cap_all = FALSE,
  split_case = TRUE,
  keep_running_caps = TRUE,
  space = "_",
  remove = NULL
)
}
\arguments{
\item{x}{A character vector}

\item{cap_all}{Logical: if \code{TRUE}, first letter of each word after splitting will be capitalized. If \code{FALSE}, only the first character of the string will be capitalized. Note that in order to balance this with respecting consecutive capital letters, such as from acronyms,}

\item{split_case}{Logical: if \code{TRUE}, consecutive lowercase-uppercase pairs will be treated as two words to be separated.}

\item{keep_running_caps}{Logical: if \code{TRUE}, consecutive uppercase letters will be kept uppercase.}

\item{space}{Character vector of characters and/or regex patterns that should be replaced with a space to separate words.}

\item{remove}{Character vector of characters and/or regex patterns that will be removed before any other operations; if \code{NULL}, nothing is removed.}
}
\value{
A character vector with each item newly formatted
}
\description{
This function combines several ways titles and names may need to be formatted. It's meant to be simple, yet flexible.
}
\details{
Examples of possible common operations include:

\itemize{
\item "TownName" --> "Town Name"
\item "town_name" --> "Town Name"
\item "town_name" --> "Town name"
\item "RegionABC" --> "Region ABC"
\item "TOWN_NAME" --> "Town Name"
}
}
\examples{
t1 <- c("GreaterNewHaven", "greater_new_haven", "GREATER_NEW_HAVEN")
clean_titles(t1, cap_all = TRUE, keep_running_caps = FALSE)

t2 <- c("Male!CollegeGraduates", "Male CollegeGraduates")
clean_titles(t2, space = c("_", "!"))

t3 <- c("Greater BPT Men", "Greater BPT Men HBP", "GreaterBPT_men", "greaterBPT")
clean_titles(t3, cap_all = FALSE)

t4 <- c("New Haven town, New Haven County, Connecticut",
        "Newtown town, Fairfield County, Connecticut")
clean_titles(t4, cap_all = TRUE, remove = " town,.+")
}
