% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_prev.R
\name{color_prev}
\alias{color_prev}
\title{Preview color palette}
\usage{
color_prev(colors, labels = T, label_color = "black", border = F,
  border_color = "white", border_size = 0.5)
}
\arguments{
\item{colors}{Character vector of colors}

\item{labels}{Logical; whether to show color name/hexcode labels}

\item{label_color}{Color to use for optional labels}

\item{border}{Logical; whether to draw border around rects}

\item{border_color}{Color to use for optional borders}

\item{border_size}{Size of optional borders}
}
\value{
\code{ggplot} object showing rects for each color
}
\description{
Preview color palette
}
