% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_label_theme.R
\name{draw_label_theme}
\alias{draw_label_theme}
\title{Create \code{cowplot} label based on a ggplot theme}
\usage{
draw_label_theme(label, theme = NULL, element = "text", ...)
}
\arguments{
\item{label}{A string of text for label}

\item{theme}{A ggplot theme; if blank, will get current theme with \code{ggplot2::theme_get}}

\item{element}{Name of a theme element; defaults to base text}

\item{...}{Any other arguments to pass to \code{cowplot::draw_label}}
}
\value{
A \code{ggplot} object
}
\description{
Create \code{cowplot} label based on a ggplot theme
}
\examples{
title <- cowplot::ggdraw() +
  draw_label_theme("Plot title", ggplot2::theme_bw(), element = "plot.title", x = 0.05)
}
