% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{edu}
\alias{edu}
\title{Education levels}
\format{A tibble with 15 rows and 4 columns
\describe{
\item{name}{Town name}
\item{variable}{Education level}
\item{estimate}{Estimated count}
\item{moe}{Margin of error of estimates}
}}
\source{
US Census Bureau, American Community Survey 2016 5-year estimates
}
\usage{
edu
}
\description{
Educational attainment data for 5 New Haven-area towns, 2016
}
\keyword{datasets}
