% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gnh}
\alias{gnh}
\title{Greater New Haven towns}
\format{
A tibble with 13 rows and 2 columns
\describe{
\item{town}{Town name}
\item{region}{Region name}
}
}
\usage{
gnh
}
\description{
Names of towns in Greater New Haven, with their region names, as defined by DataHaven
}
\keyword{datasets}
