% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_wide.R
\name{make_wide}
\alias{make_wide}
\title{Arrange estimates, MOEs, etc into wide format}
\usage{
make_wide(df, ..., group = variable)
}
\arguments{
\item{df}{A data frame}

\item{...}{Bare column names of types of measures}

\item{group}{Bare column name of group or category variable for spreading into wide data}
}
\value{
A tibble/data frame in wide format
}
\description{
Arrange estimates, MOEs, etc into wide format
}
\examples{
edu \%>\%
  make_wide(estimate, moe)
}
