% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moe_test.R
\name{moe_test}
\alias{moe_test}
\title{Significance testing of differences with MOEs}
\usage{
moe_test(df, est1, moe1, est2, moe2, cl = 0.9, alpha = 0.05,
  show_calc = T)
}
\arguments{
\item{df}{A data frame}

\item{est1}{Estimate for first group}

\item{moe1}{Margin of error for first group}

\item{est2}{Estimate for second group}

\item{moe2}{Margin of error for second group}

\item{cl}{Confidence level used in calculating MOEs given; defaults to 0.9, per ACS data}

\item{alpha}{Alpha used for significance testing; defaults to 0.05}

\item{show_calc}{Logical, whether to keep intermediary calculations (default) or only result of testing}
}
\value{
A tibble/data frame with testing-related columns added
}
\description{
Significance testing of differences with MOEs
}
\examples{
med_age <- tibble(name = c("Hamden", "New Haven"),
    men_est = c(37.2, 29.5), men_moe = c(1.9, 0.8),
    women_est = c(37.8, 31.9), women_moe = c(1.9, 0.8))
med_age \%>\%
  moe_test(men_est, men_moe, women_est, women_moe, alpha = 0.9, show_calc = T)
}
