% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pov_age_10_16}
\alias{pov_age_10_16}
\title{Ratios to poverty by age for 2 years}
\format{
A tibble with 3380 rows and 6 columns
\describe{
\item{name}{Town name}
\item{year}{Endyear of estimates}
\item{age}{Age group}
\item{ratio}{Ratio bracket}
\item{estimate}{Estimated count}
\item{moe}{Margin of error of estimates}
}
}
\source{
US Census Bureau, American Community Survey 2010 and 2016 5-year estimates
}
\usage{
pov_age_10_16
}
\description{
Population by age group by ratio of household income to the federal poverty level for Greater New Haven towns, 2010 and 2016
}
\keyword{datasets}
