% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themed_label.R
\name{themed_label}
\alias{themed_label}
\title{Create \code{cowplot} label based on a ggplot theme}
\usage{
themed_label(label, theme = NULL, element = "text", ...)
}
\arguments{
\item{label}{A string of text for label.}

\item{theme}{A ggplot theme; if \code{NULL} (the default), will get current theme with \code{ggplot2::theme_get()}.}

\item{element}{Name of a theme element; defaults to base text.}

\item{...}{Any other arguments to pass to \code{cowplot::draw_label()}.}
}
\value{
A \code{ggplot} object.
}
\description{
This is a wrapper around \code{cowplot::draw_label()} that creates a \code{ggplot}-based label that inherits formatting from a given theme element.
}
\examples{
title <- cowplot::ggdraw() +
  themed_label("Plot title", ggplot2::theme_bw(), element = "plot.title", x = 0.05)
}
