% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brk_labels.R
\name{brk_labels}
\alias{brk_labels}
\title{Create pretty breaks labels}
\usage{
brk_labels(x, format = NULL, mult_by = 1, sep = " to ")
}
\arguments{
\item{x}{A character vector}

\item{format}{A string ("percent", "dollar", "dollark") giving desired output format, or `NULL`` for no formatting}

\item{mult_by}{A number by which to multiply values in breaks}

\item{sep}{A string by which to separate values in breaks}
}
\value{
A character vector of formatted break labels
}
\description{
Create pretty breaks labels
}
\examples{
mpg \%>\%
  mutate(cty_brk = classIntervals(cty, 5)$brk \%>\% cut(cty, breaks = ., include.lowest = T)) \%>\%
  ggplot(aes(x = displ, y = hwy, color = cty_brk)) +
  geom_point() +
  scale_color_brewer(labels = brk_labels)
}
