% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brk_labels.R
\name{brk_labels}
\alias{brk_labels}
\title{Create pretty breaks labels}
\usage{
brk_labels(x, format = NULL, mult_by = 1, sep = " to ")
}
\arguments{
\item{x}{A character vector}

\item{format}{A string ("percent", "dollar", "dollark") giving desired output format, or `NULL`` for no formatting}

\item{mult_by}{A number by which to multiply values in breaks}

\item{sep}{A string by which to separate values in breaks}
}
\value{
A character vector of formatted break labels
}
\description{
Create pretty breaks labels
}
\examples{
brks <- c("[0.04,0.15]", "(0.15,0.25]", "(0.25,0.4]", "(0.4,0.55]", "(0.55,0.6]")
brk_labels(brks)
brk_labels(brks, format = "percent", mult_by = 100)
}
