% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_shares.R
\name{calc_shares}
\alias{calc_shares}
\title{Make table of rates given a denominator}
\usage{
calc_shares(df, group = group, denom = "Total", estimate = estimate,
  moe = NULL)
}
\arguments{
\item{df}{A data frame}

\item{group}{Bare column name where groups are given--that is, the denominator value should be found in this column}

\item{denom}{String; denominator to filter from \code{group}}

\item{estimate}{Bare column name of estimates or values}

\item{moe}{Bare column name of margins of error; if supplied, MOE of shares will be included in output}
}
\value{
A tibble/data frame with shares (and optionally MOE of shares) of subgroup values within a denominator group. Shares given for denominator group will be blank.
}
\description{
Make table of rates given a denominator
}
\examples{
edu \%>\%
  dplyr::group_by(name) \%>\%
  calc_shares(group = variable, denom = "age25plus", moe = moe)
}
