% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_shares.R
\name{calc_shares}
\alias{calc_shares}
\title{Make table of rates given a denominator}
\usage{
calc_shares(
  .data,
  ...,
  group = group,
  denom = "total_pop",
  value = estimate,
  moe = NULL,
  digits = 3,
  estimate = NULL
)
}
\arguments{
\item{.data}{A data frame}

\item{...}{Optional; bare column names to be used for groupings.}

\item{group}{Bare column name where groups are given--that is, the denominator value should be found in this column}

\item{denom}{String; denominator to filter from \code{group}}

\item{value}{Bare column name of values. Replaces \code{estimate} argument, but (for now) still defaults to a column named \code{estimate}}

\item{moe}{Bare column name of margins of error; if supplied, MOE of shares will be included in output}

\item{digits}{Number of digits to round to; defaults to 3.}

\item{estimate}{\emph{Soft deprecated; use} \code{value}. Bare column name of estimates or values}
}
\value{
A tibble/data frame with shares (and optionally MOE of shares) of subgroup values within a denominator group. Shares given for denominator group will be blank.
}
\description{
\code{calc_shares} makes it easy to divide values by some denominator within the same long-shaped data frame. For example, it works well for a table of population groups for multiple locations where you want to divide population counts by some total population. It optionally handles dividing margins of error. Denote locations or other groupings by using a grouped data frame, passing bare column names to \code{...}, or both.
}
\examples{
edu \%>\%
  dplyr::group_by(name) \%>\%
  calc_shares(group = variable, denom = "age25plus",
              value = estimate, moe = moe)

race_pops \%>\%
  calc_shares(name, group = variable, denom = "total",
              value = estimate, moe = moe)
}
