% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{edu}
\alias{edu}
\alias{edu}
\alias{edu_detail}
\title{Education levels}
\format{A tibble with 4 columns. For \code{edu}, 15 rows; for \code{edu_detail}, 125 rows.
\describe{
\item{name}{Town name}
\item{variable}{Education level}
\item{estimate}{Estimated count}
\item{moe}{Margin of error of estimates}
}}
\source{
US Census Bureau, American Community Survey 2016 5-year estimates
}
\usage{
edu

edu_detail
}
\description{
Educational attainment data for adults ages 25 and over for 5 New Haven-area towns, 2016. \code{edu} is an abridged version with just counts of adults with Bachelor's and Master's degrees. \code{edu_detail} contains detailed education levels from no schooling through Doctorate.
}
\keyword{datasets}
